# HEX-LoadBench/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HEX-LoadBench API
  description: |
    Load-Testing & API Stress Test Platform
    
    ## WARNING
    THIS TOOL IS FOR AUTHORIZED PERFORMANCE TESTING ONLY.
    
    Users are responsible for ensuring they have proper authorization before testing any systems.
    
    ## Authentication
    This API uses JWT tokens and API keys for authentication.
    
    - JWT tokens are used for user authentication (session-based)
    - API keys are used for service-to-service communication
    - All endpoints require authentication except `/health` and `/docs`
    
  version: 1.0.0
  contact:
    name: HEX-LoadBench Support
    email: support@hexloadbench.com
    url: https://hexloadbench.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.hexloadbench.com
    description: Production server
paths:
...
  /tests:
    post:
      tags:
        - Tests
      summary: Create test definition
      description: Create a new load test definition
      operationId: createTest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestDefinitionCreate'
      responses:
        '201':
          description: Test created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestDefinitionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tests/{test_id}:
    get:
      tags:
        - Tests
      summary: Get test definition
      description: Returns a specific test definition by ID
      operationId: getTest
      security:
        - bearerAuth: []
      parameters:
        - name: test_id
          in: path
          required: true
          description: Test ID
          schema:
            type: integer
      responses:
        '200':
          description: Test retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestDefinitionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Test not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tests/{test_id}/start:
    post:
      tags:
        - Tests
      summary: Start test execution
      description: Start executing a load test
      operationId: startTest
      security:
        - bearerAuth: []
      parameters:
        - name: test_id
          in: path
          required: true
          description: Test ID
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestStartRequest'
      responses:
        '200':
          description: Test started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestExecutionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Test not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Test is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tests/{test_id}/stop:
    post:
      tags:
        - Tests
      summary: Stop test execution
      description: Stop a running test
      operationId: stopTest
      security:
        - bearerAuth: []
      parameters:
        - name: test_id
          in: path
          required: true
          description: Test ID
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestStopRequest'
      responses:
        '200':
          description: Test stopped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestExecutionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Test not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tests/{test_id}/executions:
    get:
      tags:
        - Tests
      summary: Get test executions
      description: Returns all executions for a specific test
      operationId: getTestExecutions
      security:
        - bearerAuth: []
      parameters:
        - name: test_id
          in: path
          required: true
          description: Test ID
          schema:
            type: integer
      responses:
        '200':
          description: Executions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestExecutionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Test not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: number
          format: float
          description: Unix timestamp
...
    TestStatus:
      type: string
      enum:
        - draft
        - scheduled
        - running
        - completed
        - failed
        - aborted
        - cancelled
      description: Test status
    LoadProfileType:
      type: string
      enum:
        - ramp_up
        - steady_state
        - spike
        - soak
        - custom
      description: Type of load profile
    TestDefinitionCreate:
      type: object
      required:
        - name
        - target_url
        - load_profile_type
        - duration
      properties:
        name:
          type: string
          description: Test name
        description:
          type: string
          description: Test description
        target_url:
          type: string
          format: uri
          description: Target URL for testing
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS]
          default: GET
          description: HTTP method
        headers:
          type: object
          additionalProperties:
            type: string
          description: HTTP headers
        body_template:
          type: string
          description: Request body template
        auth_type:
          type: string
          description: Authentication type
        auth_credentials:
          type: object
          description: Authentication credentials
        load_profile_type:
          $ref: '#/components/schemas/LoadProfileType'
        load_profile_config:
          type: object
          description: Load profile configuration
        duration:
          type: integer
          description: Test duration in seconds
        max_concurrent_users:
          type: integer
          default: 100
          description: Maximum concurrent users
        rps_limit:
          type: integer
          default: 1000
          description: RPS limit
        timeout:
          type: integer
          default: 30
          description: Request timeout in seconds
        region:
          type: string
          default: us-east-1
          description: Test region
        use_proxies:
          type: boolean
          default: false
          description: Whether to use proxies
        authorization_document_hash:
          type: string
          description: Authorization document hash
        justification:
          type: string
          description: Test justification
    TestDefinitionResponse:
      type: object
      properties:
        id:
          type: integer
          description: Test ID
        name:
          type: string
          description: Test name
        description:
          type: string
          description: Test description
        target_url:
          type: string
          description: Target URL
        method:
          type: string
          description: HTTP method
        load_profile_type:
          $ref: '#/components/schemas/LoadProfileType'
        duration:
          type: integer
          description: Test duration in seconds
        max_concurrent_users:
          type: integer
          description: Maximum concurrent users
        rps_limit:
          type: integer
          description: RPS limit
        region:
          type: string
          description: Test region
        status:
          $ref: '#/components/schemas/TestStatus'
        is_active:
          type: boolean
          description: Whether the test is active
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        created_by:
          type: integer
          description: Creator user ID
    TestStartRequest:
      type: object
      properties:
        test_id:
          type: integer
          description: Test ID
        overrides:
          type: object
          description: Test configuration overrides
    TestStopRequest:
      type: object
      properties:
        test_id:
          type: integer
          description: Test ID
        reason:
          type: string
          description: Stop reason
    TestExecutionResponse:
      type: object
      properties:
        id:
          type: integer
          description: Execution ID
        execution_id:
          type: string
          description: Unique execution identifier
        test_definition_id:
          type: integer
          description: Test definition ID
        status:
          $ref: '#/components/schemas/TestStatus'
        started_at:
          type: string
          format: date-time
          description: Start timestamp
        completed_at:
          type: string
          format: date-time
          description: Completion timestamp
        duration_actual:
          type: integer
          description: Actual duration in seconds
        allocated_users:
          type: integer
          description: Allocated users
        allocated_rps:
          type: integer
          description: Allocated RPS
        runner_type:
          type: string
          description: Runner type (k6, python, go)
        runner_instance:
          type: string
          description: Runner instance ID
        error_message:
          type: string
          description: Error message
        abort_reason:
          type: string
          description: Abort reason
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
...